requirements.txt
anthropic==0.94.0
python-telegram-bot[job-queue]==22.7
google-auth-oauthlib==1.3.1
google-api-python-client==2.194.0
python-dotenv==1.2.2
pytz==2026.1.post1
ddgs==9.13.0
feedparser==6.0.12
openai==2.31.0
pypdf==6.9.2
requests==2.33.1
openpyxl==3.1.5
beautifulsoup4==4.14.3
filelock==3.28.0

# RAG (sémantique, local)
chromadb==1.5.7
sentence-transformers==5.4.1
numpy<2
torch==2.2.2
transformers==4.57.6

# Optionnel (Twilio SMS/WhatsApp, install seulement si besoin)
# twilio
